Initial codec module
