+++ beacon_node/eth2-libp2p/src/rpc/codecs/mod.rs
@@ -1,0 +1,1 @@
+mod base;
