+++ beacon_node/eth2-libp2p/src/rpc/mod.rs
@@ -16,6 +16,7 @@ use slog::o;
 use std::marker::PhantomData;
 use tokio::io::{AsyncRead, AsyncWrite};
 
+mod codecs;
 mod handler;
 pub mod methods;
 mod protocol;
